Client::Sync::*::testItems: check that items are added on server
The Oracle server has a problem where it accepts ical20 items,
then tells the client to delete them. This happened to pass the
test because client-test then compared an empty client A against
and empty client B.

Now the item counts are checked. The expectation are that no changes
are made except for adding an unknown number of items to the server.