Oracle: tentative workaround for UID+RECURRENCE-ID problem
The Oracle Beehive server
a) does not support RECURRENCE-ID and
b) expects that UID is different between all VEVENTs,
even those belonging to the same recurring event.

Achieve that by attaching the RECURRENCE-ID (= ORIGSTART)
to the UID and not sending it separately.

Demangling such a UID is missing and cannot be tested right now
because the server deletes all items after receiving them.