SyncContext: abort before sending message
A good point to abort a sync is after the engine has processed
a message and before sending out the reply. This patch adds that
check for STEPCMD_SENDDATA and checkForAbort() directly after
the SessionStep() call.

Previously, the transport agent would be invoked. Eventually that
transmission also gets cancelled once the transport checks for
abort, but doing it sooner is better and more obvious.

Because teh whole suspend/abort handling is somewhat tricky, this
patch adds debug logging around those state changes.