Synthesis plugin: static linking without --export-dynamic
The previous solution with lookup of "SyncEvolution_<sym>" via
dlsym(RTLD_DEFAULT) in libsynthesis was non-portable. It depends on
the visibility of these functions and thus --export-dynamic.

The --export-dynamic flag is GNU ld specific. It doesn't work
on Mac OS X (unknown option) and on Android (accepted, but had no
effect).

This patch solves this by specifying the plugin functions directly
via the //static/<sym>=<addr>/... syntax. This only works with
a libsynthesis that supports that.

The advantage even on systems which support --export-dynamic is
the smaller function table and thus faster loading of the executable.
The disadvantage is the requirement to add new plugin functions
to the internal list of functions.