Html UI: add framework for HTML UI
It is possible to use HTML/Javascript to write a portable
UI with the help of browser communicating with dbus.
Browser dbus bridge is a open source project which provides
a bridge to browser and dbus. Thus we can write UI to access
syncevo-dbus-server.
Currently the ui.html could be opened in the browser and do
synchronizations as required. It is a simple sample now.
The more detailed infomartion is to be added in future.
The build steps are:
1) change directory to src/htmlui/
2) make xpcom
3) sudo make install
How to use it:
open the html file 'src/html/ui/ui.html' in the firefox.