syncevo-phone-config: reimplemented template generation and calendar+todo detection (MBC #1197)
--combined-calendar-task was broken: because "calendar+todo" was
tested with only an event, the test always passed when using the
settings for events, even if tasks cannot be exchanged using those
settings. If it had worked and had been essential to get correct
results, the option should have been enabled by default.

The new implementation of combined calendar+todo storage works
differently, without increasing the number of tests that need to
be run when searching for complete settings:
- detect settings for "calendar" using VEVENT as test data
- detect settings for "todo" using VTODO as test data
- if uri identical, configure a "calendar+todo" source instead of
  separate "calendar" and "todo"

Note that this hasn't been tested yet.

The --combined-calendar-task option is thus obsolete and was removed.

The template generation was rewritten and now prints the template
to stdout. While doing this, some more changes were made:
- "type" is always set, even if using default data formats,
  because "default" data formats are a bit underspecified and it
  is better to be explicit about what worked
- previously, "type" did not distinguish between events and tasks
  because it omitted the "calendar:" and "todo:" type part
- the template is printed if and only if --advanced is enabled,
  because only then is the result reliable; disabling it makes
  little sense when printing to stdout, so --create-template was
  removed

The generated config is no longer derived from the 'Nokia 7210c'
template. Instead the "SyncEvolution Client" template is used, because
it does not contain a potentially unused "calendar+todo" source.
The same fix for missing "calendar/todo" in "type" also had to
be applied to to the created config.

The syncevo-phone-config tool used source names that differed from the
syncevolution command line, for example "contact" was used instead of
"addressbook". This was not only making the code more complex (it had
two different ways to map "contact" to "addressbook"), it was also
exposed to users via the --source option. Now the same
addressbook/calendar/memo/todo names as in the rest of the
SyncEvolution command line are used.

This patch also includes quite a bit of code, language and output
cleanup. The code still isn't as pythonic as it could be, though.