syncevo-phone-config: fix combined calendar detection for Nokia phones?! (MBC #1197)
After thinking some more about this, I think the current version of
the script will fail to detect combined calendar+todo for Nokias,
because of the changed order of URIs that we test. I think Nokia
phones ignore the URI that we send to it and then decides on the data
store entirely based on the data type. That way the first URIs for
calendar and todo are accepted ("Calendar" and "Task"), which breaks
the "URI equal" comparison.

The fix should be simple - test the calendar URIs first when testing
todos, as it was done before my last commit.