file backend: added "raw" mode
The "raw" mode is selected by forcing the type with a trailing
exclamation mark, for example:
   type=file:text/vcard:3.0!

In raw mode, items are not converted by the Synthesis engine. This is
the behavior of SyncEvolution's file backend in releases < 0.9. It
is useful when treating the file backend as the local backup of
a SyncML server. It is not useful when using it as a SyncML server,
because that depends on operations not supported in raw mode.

The syntax with exclamation mark was valid before, but had no effect
because a file sync source always only supported. The implementation
of the raw mode depends on the "raw" items in the Synthesis engine,
which is selected by defining a field list with a single "ITEMDATA"
field, then defining data types with basetype="raw" using it.

This patch also renames the local "data" variable into "itemdata" so
that the SyncSourceSerialize code works with and without raw mode.