SyncML server progress: sync mode unknown (MBC #2786)
Because PEV_ALERTED is not generated, SyncEvolution doesn't record a
"final sync mode" for the started sources. One of the effects is that
intermediate progress events (item processed/sent/received) were not
printed.

This patch address this part of the problem by introducing a new
SYNC_UNKNOWN sync mode. It is set once it is known that a source is
part of the sync (PEV_STARTED) and remains set unless overwritten with
a PEV_ALERTED later on (as in clients).

This is an intermediate solution. For example, statistics now say
"unknown" instead of "two-way/refresh-from-server/...", which is not
what the user understands easily and wants. The real solution is to
determine the sync mode in which a source operates also in server
mode.