akonadisyncsource.cpp: Fixed the bug where the remote changes aren't being shown in local database
The remote changes weren't being shown in the local database, even though the sync was reported successful.
On the next sync, the remote changes are being wiped off. This was happening because of an improper item modify job.
Still need to take care of the TODO in the same function.