buteo: fix issue about buteo retrieving deleteditems
Buteo uses fixed database to record deleted items for contacts.
It tracks 'delete' information by querying this one.
As a result, incorrect 'delete' records are used to sync for two
clients by sharing the same one.

To fix this, we create different database for each one. The idea is
to clean them before running a test. Then at the first time the db is used,
we copy default database and generating by wiping out all records of
default one.