dbus-server: Moved all class definitions to .h file.
Additionally, I've moved the includes into a common.h file. I'll keep it this way till the classes are broken out into there own files.

There are a few global variables that were static. I've removed the static keyword for the time being and have marked them with a FIXME.