a autotools: fixed static linking issues in previous commit
The static library case in the previous commit fails because the
object files in the libsynccommon.la convenience library are not
pulled into the executables.

Adding --whole-archive as in this commit fixes the problem, but is
only a bad hack at best (must trick libtool into preserving the
ordering, makes assumptions about libtool internal .libs).