server: Check if vendor id was assigned by Bluetooth SIG
Attribute ID 0x0205 of the PnPInformation service record tells us if the vendor id was assigned by the Bluetooth SIG or the USB Implementor's group.

This patch checks that the Vendor ID was assigned by the Bluetooth SIG
and ignores the PnPInformation if it was not. It is unlikely that
mobile phone manufacturers have vendor IDs not assigned by the
Bluetooth SIG. Regardless, eventually we should also use the vendor
ID's assigned by the USB Implementor's group as well.

For a list of USB assigned vendor id's see
http://www.linux-usb.org/usb.ids