tests: Added 2 very basic tests for Bluetooth devices
TestBluetooth.testBluetoothProductId tests the product id is the same
as in the lookup table. You'll want to modify the name on the test
device to something other than what's in the lookup table.

TestBluetooth.testBluetoothTemplates tests that at least the first
config template is present in the list returned by GetConfigs for the
device being tested.

These tests are disabled by default.