Use PnPInformation info also when no product id matches.
The previous commit would only use the PnpInformation information if
both the vendor and product id matched.

This commit makes it so that if a vendor id is found but no product
id, the product name is set to an empty string. We want this because
we can still use the vendor information without a product id.