testing: improved Client::Source::*::testImport
Instead of aborting the test at the first item which
fails to import, collect the error messages (from the
exceptions), do the comparison of imported data
against the reference data and then either report
the import failures (if there were any) or the comparison
failure (if that broke).

The advantage is that a single run of the test shows
all problems that exist.