syncevo: Add m_peerName to TemplateDescription
When a bluetooth device has a PnpInformation service record, the
user-modifiable device name was no longer available through the d-bus
interface. Previously the fingerprint was used to store and retrieve
this value.

This commit creates a m_peerName field in TemplateDescription to hold
this value. A peerName parameter was added to the TemplateDescription
constructor and the value of DeviceDescription.m_deviceName passed in.