server: Place bluetooth product look-up table in own file
The lookup table needs to be in a seperate file so that distros and
users can add their own products without recomiling.

This commit moves the bluetooth lookup table for products, and vendors
for the time being, into its own file. The file is in standard .ini
format as understood by GKeyFile. It uses a static variable to hold
the GKeyFile in order to prevent the file from being loaded more than
once.