engine: tell peers about detached recurrence exceptions
The Maemo 5 calendar backend does not support UID/RECURRENCE-ID
semantic. This is problematic in combination with peers which rely on
that to override the main recurring event on specific recurrences,
because the Maemo 5 calendar will show duplicates.

Fixing this inside the calendar backend is hard because it doesn't get
to see all related events at once. The engine has the same
problem. Therefore a workaround is used:
- backends which have that unified access to all related events
  add special X-SYNCEVOLUTION-EXDATE-DETACHED properties to the
  main event which correspond to the exceptions represented by
  detached recurrences, in addition to the regular EXDATE; the
  backend must set the HAVE-SYNCEVOLUTION-EXDATE-DETACHED rule when
  using PARSETEXTWITHPROFILE()
- the engine parses the special property (only if the
  HAVE-SYNCEVOLUTION-EXDATE-DETACHED rule is active!) and stores
  the values in the EXDATES_DETACHED field
- this gets passed between SyncEvolution instances in a local
  or SyncML sync
- another SyncEvolution backend can request to get either
  X-SYNCEVOLUTION-EXDATE-DETACHED or regular EXDATE properties
  by setting the HAVE-[SYNCEVOLUTION-]EXDATE-DETACHED rules

The Maemo 5 backend needs to set HAVE-EXDATE-DETACHED. This means that
importing events will add additional EXDATEs which will be sent back
when updating events inside the calendar app, but that is okay. They
are merely redundant.

(cherry picked from commit a4d0909b28721ff3500306f5a29a367ba2f2bffb)

Conflicts:

	src/syncevo/MapSyncSource.h