More changes for gio gdbus wrapper. Now compiles.
These changes get us to the point where we can compile. It doesn't
work yet, though.

One of the tricky things to note is the different way method handling
is dealt with. I needed to create a struct, MethodHandler, to deal
with this. The static handler method is called whenever the dbus
objects has one of its methods called. This then checks if the method
has been registered and calls the appropriate callback. The std::map
which holds the method name to callback is not very nice.