CalDAV: filter out X-SYNCEVOLUTION-EXDATE-DETACHED
If the engine got a parent event with X-SYNCEVOLUTION-EXDATE-DETACHED,
merged it internally and then wrote it back, the
X-SYNCEVOLUTION-EXDATE-DETACHED would have been stored in the CalDAV
server. Now this is avoided by removing all such properties before
storing the new or updated event.

This was previously done (and still is) as an extra precaution in the
code which adds the properties.

(cherry picked from commit ede6e65ccb7cf4180711a56c6a28c4914b25a7dd)