bluez-gdbus: Make pointer conversion explicit to assist compiler
On Debian stable, the compiler (gcc 4.4.5) was not properly casting
the void *, memory, to the WatchData *, watch_data. Under gdb the
result of "print (WatchData*)memory->source provided a valid pointer
but the result of "print watch_data->source" always reulted in an
invalid pointer like 0x08.

This was causing the syncevo-dbus-server to segfault on start up.