fork/exec: added onFailure
Methods called from glib C code must not throw exceptions. If that
happens, the problem will be reported via onFailure.

The more likely problem is that the child quits prematurely or with an
error code. Both is reported via onFailure now, so the user of
ForkExecParent no longer has to watch the child itself.