Another step towards splitting the client resource into seperate binary
It currently compiles the server but only with stubs. The server does
nothing but create a dbus interface and immediately exit.

This commit focused on stripping out code that will not be needed when
concurrent syncs are a reality. The work queue has been turned into a
session list, session priority will no longer be needed, etc. At this
point the simple assumtion is made that a session helper is launched
immediately and the session will be detached when it is done.

Note: The difficult problem of sync sessions with different sources
using the same backend database is being ignored at this point.