Implement functionality to wait for dbus method callbacks
When a DBus method call is made by a client, often the server process
must make a DBus method call to the helper. The server must block when
making this call until the callback has been called and return values
are available.

This blocking mechanism is implemented by calling waitForReply() which
polls the dbus connection for activity and does a context iteration if
an error has occured.

In syncevolution.cpp, a similar block is implemented by running and
quiting the main loop. However, in the server this will not work as
we've already got a main loop running.