Implement DBus interface for Connection object
The helper binary can start a Session directly or a Connection. This
commit implements the DBus interface in the helper and corrosponding
DBusClientCalls and SignalWatches in the ConnectionResource.

This builds but will probably not yet work properly. For example, 2 of
the arguments to the Connection constructor have been removed which
require. This are to be added back in subsequent commits.

In addition, the helper was modified to accept a commandline argument,
'-c', which specifies that a Connection should be started rather than
a Session.