Move forkexec out of constructor and wait for dbus initialization
When running tests, the dbus interface was not immediately available
after calleing startSession on the server. This was because the
onConnect callback which activates the interface had not yet been
called.

This patch insures that a SessionResource is not returned unless the
Session DBus interface is ready to be consumed. The helper
initialization is moved into an init() method which is called in the
createSessionResource static method.