Add ForkExecChild::getEnvVar to retrieve environment varables in child
The commit also adds the ability to add and get environment variables
with empty values. When an empty value is add the constant SE_EMPTY is
used for the value. When this value is found while getting a value an
empty string is returned. An empty string is also returned when the
environment variable is not found.