SyncSourceReport: added "number of restarts"
This will become relevant once a SyncEvolution session may involve
multiple SyncML sessions. The "restart" counter is normally zero (for
compatibility with older sync reports on disk) and incremented once
for each time that a sync source was involved in a SyncML session after
the initial one.

The sync mode is still only stored once. It is expected to be the
one from the initial sync.