SyncSource + SyncContext: added requestAnotherSync()
The long term goal is that using a specific sync source once more in
another, new sync session can be requested per source.

The actual implementation however is per-session, using the new
libsynthesis "restartsync" session variable. This functionality
sits in SyncContext, because it is the one which has access to
the session variables.