Implement password request from helper session
Before the work to do concurrent syncs the session would make a
request and wait a specified duration till the password was either
received or the request timed out.

With the introduction of concurrent sessions and the helper binary,
the way this functions has changed a bit. When the request is made the
helper emits a password request signal and waits. The helper only does
one job, so this should be fine. The SessionResource in the server
process receives this signal and creates an InfoRequest. We can't wait
in the server process because we've got possibly multiple concurrent
sessions running. In order for the user of the InfoReq object to be
notified that a client has provided the info or the request timed out
a signal was introduced in InfoReq that lets us know when the response
is received. This signal is connected to a routine in the
SessionResource which passes the result to the session helper who can
now carry on its business.