Implement new simplified shutdown mechanism
Before the work to do concurrent sync sessions, a high priority
shutdown session was placed into the queue and started a shutdown
whenever it was activated.

This commit gits rid of the ShutdownSession as it no longer makes
sense in the new concurrent session world.

It also allows the helper session to continue its work when the server
process shuts down.  Before shutting down the serverShutdown method is
called on each session helper to let it know the server will no longer
be running.

A resulting of the above change is that the server can now exit
without waiting for all sessions to finish.

The above mentioned changes result in much simplified shutdown
functionality.

Note: This is untested!