Client::Source testing: cover multiple sync cycles
A source is expected to cope with beginSync() being called
again after endSync(). This is necessary for restarting a sync
in the same session to send further changes.

The new Client::Source::*::testChangesMultiCycles() covers this
aspect. It uses the same revised and extended doChanges()
implementation as the older testChanges() test, except that source B
is only instantiated once and then restarted instead of recreating it.