Add helper session control logic to server process
The Server class now has a queue for holding waiting resources
(Sessions or Connections) and a list for holding active resources. The
queue is actually a std::set with the Compare object initialized to
PriorityCompare, a nested class of Server.

The logic for queueing and removing resources has been reworked to
reflect these new structures.

Currenly only one resource will be active because the method for
checking that resources can run concurrently always returns
false. This is due to a yet unsolved issue in detecting if 2 session
could be accessing the same sources.