client-test: add Client::Sync::*::testTwoWayRestart
This is the first of several tests which run a sync, make changes to
the local data after the source has checked for changes as part of
that sync, then requests the sync to continue with a second cycle.

The test checks the final sync mode (same as original cycle), all
intermediate reports (captured via the endDataWrite slot) and total
number of cycles.