syncevo-dbus-server: use call operator for D-Bus calls
The blocking call operators are a lot easier to use.

Note that this is still not the right way of doing this! The
conversion is only done to get rid of the block() method calls.

The right approach is to relay messages and replies totally
asynchronously. syncevo-dbus-server must never block. That code must
also relay error messages, including the type, back to the original
caller.