testing: restart tests for all five sync modes
The previous twoWayRestart test was broken: it stored the intermediate
SyncSourceReport at the time of the first EndDataWrite, but at that
point the report hasn't been updated yet. It added the item before the
first StartDataRead, and thus the change was included in the first
cycle, not in the second one as intended.

Now the reports are stored before each StartDataRead, which is
guaranteed to include anything which has happened before. The data
modification and restart request is done after the initial
StartDataRead, at which point the source alread did its change
detection and thus doesn't include that change.

refresh-from-remote is problematic. Should it try to wipe out all
items added after the initial refresh-from-remote? The
one-way-from-remote sync that happens in later cycles doesn't do
that. The test currently expects that the item doesn't get deleted
and thus reflects the current implementation.