Pass shutdownRequested flag to Session and Connection in helper
Previously, there was know way to detect in a Session on Connection
whether a shutdown signal had been called.

This commit adds a shutdownRequested flag that the helper process
passes to the Session or Connection which can be used to detect that a
shutdown signal has been recieved. This immediately useful to exit a
busy loop while waiting for a password response.