testing: extended restart tests
Now (almost) all restart tests do a sequence of "add/update/remove
item in second cycle" syncs. For refresh-from-remote only adding can
be tested, because any further syncs remove that item as it never
reaches the server and will be deleted locally.

These changes showed that modifying data after the StartDataRead
wasn't a good idea, as it affects the first cycle (data sent to server
is the modified item, for example). Now that is done after the first
EndDataWrite.