Added a status handler and connecting function to Resource.
The status handler (printStatus) replaces the genericErrorHandler
function. It is static method, so we don't have to track Resource
instance when setting this method as a callback of asynchronous
DBus call. Also, got rid of genericCall methods - they were using
blocking DBus calls which are a no-no in dbus server. One of the
following commits will convert the code using it into asynchronous
calls with delayed responsing by using ProxyCallbacks introduced
in one of previous commits. Also, to save typing, defaultConnectTo*
methods are added to be used when all we want is just a note whether
call failed or succeeded.