Make connection resource fully asynchronous.
Some of the work here was already done by Patrick Ohly. Connection
resource creation is now asynchronous - server which wants to create
a connection has to use createConnectionResource and pass a callback
to it which will take ownership of passed resource. Such resource
is ready to use as dbus-helper is already spawned and resource's
DBus interface is activated. Also, made some cleanup of header file -
all public methods and members are put at the top of the class and
private at the bottom.