Make session resource fully asynchronous.
All of relaying DBus methods now use ProxyCallback to delay the
response until relayed message will be responeded from dbus-helper.
Session resource creation is now asynchronous - server which wants
to create a session has to use createSessionResource and pass
a callback to it which will take ownership of passed resource. Such
resource is ready to use as dbus-helper is already spawned and resource's
DBus interface is activated. Also, made some cleanup of header file -
all public methods and members are put at the top of the class and
private at the bottom.