client-test: make restart tests pass for normal peers
Introduced CLIENT_TEST_PEER_CAN_RESTART, which must be set if and only
if the peer can restart a sync. When set, the full set of restart
tests is enabled and expected to restart. Otherwise, only
testTwoWaySync is tested and expected to not do a restart despite the
request to do so.

Furthermore, when acting as server no restart tests are done because
that doesn't work in server mode.