dbus-server: Bump GIO version to 2.30 due to bugs in earlier versions
We ran into some race conditions in version 2.28 of GIO GDBus. These
were known issues that were fixed in following bug reports.

https://bugzilla.gnome.org/show_bug.cgi?id=665211
https://bugzilla.gnome.org/show_bug.cgi?id=661689

Version 2.30.2 is currently in Weezy, the next Debian stable, which I
think is the target for dropping support for the libdbus. Thus, this
version requirement bump should not disturb that plan.