syncevo: Added method to add environment variables to ForkExec
The parent may need to pass additional environment variables to the
child process.

This commit adds an AddEnvVar method to the parent process. If no
value is provided an arbitrary one is assigned. This is interesting
when the existance of the variable is of importance rather than the
associated value.