dbus-server: Run sync sessions in separate processes
This commit separates sync sessions into their own processes. Thus,
ground work is laid for running multiple sync session concurrently.

Sessions are separated into Session and SessionImpl classes. A Session
instance resides in the server process and serves as a proxy to the
SessionImpl instance which is in the child process. These two classes
communicate using the Session's DBus interface. The same relationship
is shared by Connection and ConnectionImpl classes.

In order to relay message requests made by clients communicating with
the Server to the SessionImpls running in a separate process async DBus
messages are used. This ensures that the server process does not block
while waiting for the SessionImpl to return a response.

Note: Autosync and Connections are not yet working under this scheme.