D-Bus server: prevent new sessions and connections when shutting down
When syncevo-dbus-server is preparing to shut down, it should not
allow new resources (sessions and connections) to register
themselves (because they are not meant to run) and should not run
them (happened inside addResource()).

This is achieved by throwing a "server shutting down" exception in the
D-Bus API implementation (to avoid the entire resource creation early)
and in the addSource() implementation (to catch resource where
creation was already kicked of before shutdown started).

However, throwing exections in addSource() currently kill
syncevo-dbus-server. That needs to be fixed, see "[SyncEvolution]
GDBus C++ add_filter + fork/exec dbus-server".