ForkExec: fix race condition
Delay message processing in the child until m_onConnect is done.

Fixes the following problem with GIO D-Bus:
- client connects, starts helper thread
- parent sends method call
- helper thread processes method calls, discards it
  (doesn't match anything)
- client calls m_onConnect (too late)

The introduction of delayed message processing broke the
dbus-client-server, because it does its main work in the onConnect
signal, where replies by the server where not processed after the
change.

The solution is to explicitly allow message processing inside the
signal slot.