ForkExec: make stop() and kill() NOPs when child isn't running
syncevo-dbus-server was killing pid == 0 == itself when destructing a
Session where the child couldn't be started. Check for m_childPID == 0
and don't do anything in that case.

Likewise, sending a signal to the old pid when the child has already
quit is better avoided, because the pid might already have been
recycled by the kernel for a different process (unlikely, but
possible).