D-Bus server: code cleanup for initial logger
Replaced "cast via function wrapper" with static_cast. Needed to
keep the ? : operator happy.

While looking at the code I also noted that it might be useful
to stack the syslog logger on top of the redirect logger, to
get the error logging from glib etc. into the syslog. To be decided
later...