D-Bus server: handle syncevo-dbus-helper output
This is a first shot at making the user-visible output created during
operations visible to the user again. It's based on the same idea as
output handling in the syncevo-dbus-server:
- Session registers itself as the top-most logger and sends
  SyncEvolution logging via D-Bus to the parent, which re-sends
  it with the right D-Bus object path as output of the session.
- Output redirection catches all other output and feeds it back
  to the Session log handler, from where it goes via D-Bus
  to the parent.

The advantage of this approach is that level information is made
available directly to the parent and that message boundaries are
preserved properly.

The disadvantage is that the current solution is racy: depending on
the order in which events are processed, the command line client quits
before it has printed all output from the helper. This needs further
work...

A better solution might be to capture stdout and stderr in
ForkExec.cpp, translate it back into messages and relay that to the
command line client. That would be guaranteed to capture everything
happening inside the helper.