command line: fixed error handling for invalid property assignments
xyz=1 triggered an error message with (null) inside instead of a
property name. The name is prominently given in the assignment
parameter, so just drop the explicit property name.

=1 triggered an error message with two single quotes instead of
one (because args are already passed with quotation marks) and then
crashed because of a missing return.

Also added unit tests...