logging (base, stdout and syslog): code refactoring
LoggerSyslog copied quite a bit of code from LoggerStdout. Moved that
code into the LoggerBase. That avoids code duplication and allows
LoggerSyslog to benefit from improvements made recently in that
code (prefix in each line).

As it was before, LoggerSyslog did not handle line breaks inside the
printed text. syslog() showed them as #012. Now each line is passed to
syslog() separately, with a proper prefix.

The level in the syslog() also seemed to be wrong earlier. It needs to
be based on the log level of the current message, not on the threshold
set for logging in the logger.