command line + local sync: fixed erroneous "Comparison impossible" output.
"Comparison impossible" was incorrectly printed after a successful
comparison on the target side of local sync.

The reason is that SIGCHLD SIG_IGN prevented reading the synccompare
return code. That code was probably added to prevent zombies when
syncevo-local-sync dies before it reads the return code of its
children.

Removing the SIGCHLD SIG_IGN fixes synccompare output issue.  A test
for the sync output will be committed next - the issue was found while
writing the test.

Code for handling zombies still needs to be added to all long-running
process (client-test, syncevo-dbus-server).